Increase data processor separatator count

+++ src/libs/data-processor.h
@@ -41,7 +41,7 @@ void data_processor_deinit();
 
 ProcessingState* data_processor_get_global(void);
 
-uint8_t data_processor_count(ProcessingState* state);
+uint16_t data_processor_count(ProcessingState* state);
 char* data_processor_get_string(ProcessingState* state);
 bool data_processor_get_bool(ProcessingState* state);
 int data_processor_get_int(ProcessingState* state);
